바우처 도메인 패키지 정리
